+++ backend/src/rules/rules.service.ts
@@ -32,8 +32,32 @@ export class RulesService {
           });
         }
       }
-          
-      async create(x: GradeRules): Promise<GradeRules> {
+
+      async findOne(id: number): Promise<GradeRules> {
+        try {
+            const x = await this.service.createQueryBuilder("grade_rules")
+            .where("grade_rules.id = :id", {id: id})
+            .getOne();
+            if (!x) {
+              return null;
+            }
+            let it = new GradeRules();
+            it.id = x.id;
+            it.task_id = x.task_id;
+            it.type_id = x.type_id;
+            it.grade_id = x.grade_id;
+            it.total_num = x.total_num;
+            return it;
+        } catch (error) {
+            console.error(error);
+            throw new InternalServerErrorException({
+                status: HttpStatus.BAD_REQUEST,
+                error: error
+            });
+        }
+     }
+           
+     async create(x: GradeRules): Promise<GradeRules> {
         try {
           const y = await this.service.createQueryBuilder("grade_rules")
           .insert()
@@ -57,13 +81,17 @@ export class RulesService {
         }
       }
 
-      async delete(id: number) {
+      async delete(id: number): Promise<GradeRules> {
         try {
+          const r = this.findOne(id);
+          if (r) {
             await this.service.createQueryBuilder("grade_rules")
             .delete()
             .from(grade_rules)
             .where("grade_rules.id = :id", {id: id})
             .execute();
+          }
+          return r;
         } catch (error) {
             console.error(error);
             throw new InternalServerErrorException({
