+++ backend/src/schedules/schedules.service.ts
@@ -33,6 +33,29 @@ export class SchedulesService {
         }
     }
 
+    async findOne(id: number): Promise<ScheduleType> {
+        try {
+            const x = await this.service.createQueryBuilder("group_tasks")
+            .where("group_tasks.id = :id", {id: id})
+            .getOne();
+            if (!x) {
+              return null;
+            }
+            let it = new ScheduleType();
+            it.id = x.id;
+            it.task_id = x.task_id;
+            it.group_id = x.group_id;
+            it.created = x.created;
+            return it;
+        } catch (error) {
+            console.error(error);
+            throw new InternalServerErrorException({
+                status: HttpStatus.BAD_REQUEST,
+                error: error
+            });
+        }
+    }
+    
     async create(x: ScheduleType): Promise<ScheduleType> {
         try {
             const y = await this.service.createQueryBuilder("group_tasks")
@@ -56,13 +79,17 @@ export class SchedulesService {
         }
     }
 
-    async delete(id: number) {
+    async delete(id: number): Promise<ScheduleType>  {
         try {
-            await this.service.createQueryBuilder("group_tasks")
-            .delete()
-            .from(group_tasks)
-            .where("group_tasks.id = :id", {id: id})
-            .execute();
+            const r = this.findOne(id);
+            if (r) {
+                await this.service.createQueryBuilder("group_tasks")
+                .delete()
+                .from(group_tasks)
+                .where("group_tasks.id = :id", {id: id})
+                .execute();
+            }
+            return r;
         } catch (error) {
             console.error(error);
             throw new InternalServerErrorException({
